Add battery conversion to percentage

+++ health-check/adapter/states.go
@@ -8,7 +8,7 @@ type States struct {
 type NodeState struct {
 	NodeID   int  	   `json:"nid"`
 	State    bool	   `json:"state"`
-	Battery  int 	   `json:"battery"`
+	Battery  float64   `json:"battery"` // 전압 (14 ~ 16.8)
 	Location []float64 `json:"location"` // [lat, lon, alt]
 }
 type HealthInfo struct {
